changes based on code review by JP and Ben
